pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "little-tree-wallpaper-refactor"
version = "2.0.0"
description = "PyWebview + MUI desktop wallpaper manager inspired by Little Tree Wallpaper Next"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.11.18",
    "filetype>=1.2.0",
    "littledl>=0.4.1",
    "loguru>=0.7.3",
    "ltws-parser>=1.5.1",
    "orjson>=3.10.18",
    "pillow>=11.2.1",
    "pillow-avif-plugin>=1.5.2",
    "platformdirs>=4.3.6",
    "psutil>=7.0.0",
    "pyperclip>=1.9.0",
    "pystray>=0.19.5",
    "pywebview>=5.1",
    "requests>=2.32.3",
    "rtoml>=0.12.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
